On Radeon open-source drivers on Linux (Gallium 0.4 on ATI RV530), stricter rules for converting int->float in GLSL must be followed.
git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle_game_engine@11305 afe8dcc5-9094-451b-8773-5f495042a737

+++ src/x3d/opengl/glsl/ssao.glsl
@@ -67,9 +67,9 @@ void main(void)
     fade *= 0.5;
     for (int j = 0 ; j < samples*i; j += 1)
     {
-      float step = PI*2.0 / (samples*i);
-      pw = (cos(j*step)*i);
-      ph = (sin(j*step)*i);
+      float step = PI*2.0 / float(samples*i);
+      pw = (cos(float(j)*step)*float(i));
+      ph = (sin(float(j)*step)*float(i));
       d = readDepth(vec2(texCoord.s+pw*w,texCoord.t+ph*h));
       ao += compareDepths(depth,d)*fade;
       s += 1.0*fade;
